+++ lib/nova/api/search_params/apportionment.rb
@@ -3,7 +3,7 @@ module Nova
     module SearchParams
       class Apportionment < Nova::API::Utils::BaseStruct
         attribute? :q, Dry::Types['coercible.string'].optional
-        attribute? :with_deleted, Dry::Types['strict.bool'].optional
+        attribute? :with_inactive, Dry::Types['strict.bool'].optional
       end
     end
   end