+++ lib/nova/api/search_params/current_asset.rb
@@ -3,7 +3,7 @@ module Nova
     module SearchParams
       class CurrentAsset < Nova::API::Utils::BaseStruct
         attribute? :company_id, Dry::Types['coercible.integer'].optional
-        attribute? :with_deleted, Dry::Types['strict.bool'].optional
+        attribute? :with_inactive, Dry::Types['strict.bool'].optional
       end
     end
   end