+++ spec/nova/api/search_params/apportionment_spec.rb
@@ -3,6 +3,6 @@ RSpec.describe Nova::API::SearchParams::Apportionment do
     subject { described_class }
 
     it { is_expected.to have_attribute(:q, Dry::Types['coercible.string'].optional) }
-    it { is_expected.to have_attribute(:with_deleted, Dry::Types['strict.bool'].optional) }
+    it { is_expected.to have_attribute(:with_inactive, Dry::Types['strict.bool'].optional) }
   end
 end