+++ spec/nova/api/search_params/current_asset_spec.rb
@@ -3,6 +3,6 @@ RSpec.describe Nova::API::SearchParams::CurrentAsset do
     subject { described_class }
 
     it { is_expected.to have_attribute(:company_id, Dry::Types['coercible.integer'].optional) }
-    it { is_expected.to have_attribute(:with_deleted, Dry::Types['strict.bool'].optional) }
+    it { is_expected.to have_attribute(:with_inactive, Dry::Types['strict.bool'].optional) }
   end
 end